clean: Move argument parsing to separate file
